requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
celery
redis
pydantic
email-validator
httpx